Move CasOAuthThrottleWebMvcConfigurer inside Configuration
