+++ support/cas-server-support-oauth/src/main/java/org/apereo/cas/config/CasOAuthThrottleConfiguration.java
@@ -17,10 +17,14 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import static org.apereo.cas.support.oauth.OAuth20Constants.BASE_OAUTH20_URL;
 
 /**
  * This is {@link CasOAuthThrottleConfiguration}.
@@ -32,6 +36,20 @@ import java.util.stream.Stream;
 @Configuration("oauthThrottleConfiguration")
 @EnableConfigurationProperties(CasConfigurationProperties.class)
 public class CasOAuthThrottleConfiguration implements AuthenticationThrottlingExecutionPlanConfigurer {
+
+    @Configuration("oauthThrottleWebMvcConfigurer")
+    static class CasOAuthThrottleWebMvcConfigurer extends WebMvcConfigurerAdapter {
+
+        @Autowired
+        @Qualifier("authenticationThrottlingExecutionPlan")
+        private AuthenticationThrottlingExecutionPlan authenticationThrottlingExecutionPlan;
+
+        @Override
+        public void addInterceptors(final InterceptorRegistry registry) {
+            authenticationThrottlingExecutionPlan.getAuthenticationThrottleInterceptors().forEach(handler ->
+                    registry.addInterceptor(handler).addPathPatterns(BASE_OAUTH20_URL.concat("/").concat("*")));
+        }
+    }
 
     @Autowired
     @Qualifier("oauthSecConfig")
