+++ support/cas-server-support-oauth/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.config.CasOAuthConfiguration,\
   org.apereo.cas.config.CasOAuthAuthenticationServiceSelectionStrategyConfiguration,\
   org.apereo.cas.config.CasOAuthComponentSerializationConfiguration,\
-  org.apereo.cas.config.CasOAuthThrottleConfiguration,\
-  org.apereo.cas.config.CasOAuthThrottleWebMvcConfigurer
+  org.apereo.cas.config.CasOAuthThrottleConfiguration
